+++ zelda-service/src/main/java/com/projetofinal/zeldaservice/ZeldaServiceApplication.java
@@ -1,0 +1,15 @@
+package com.projetofinal.zeldaservice;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ZeldaServiceApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(ZeldaServiceApplication.class, args);
+
+
+	}
+
+}
