+++ zelda-service/src/main/java/com/projetofinal/zeldaservice/model/RetornoJogoPorId.java
@@ -1,6 +1,5 @@
 package com.projetofinal.zeldaservice.model;
 
-import com.projetofinal.zeldaservice.Zelda;
 import lombok.Data;
 
 @Data
