+++ zelda-service/src/main/java/com/projetofinal/zeldaservice/model/Zelda.java
@@ -1,4 +1,4 @@
-package com.projetofinal.zeldaservice;
+package com.projetofinal.zeldaservice.model;
 
 import lombok.Data;
